cleaning directory
