--- python-port/Package/README.md
@@ -1,26 +1,0 @@
-# Python3 port
-
-This is the Python port of the already excellent Bash Script.
-
-I will use the PyPi package manager to upload this.
-
-# Features
-
-The features included in this script will not likely have all the 
-features contained in the bash script. However, the following features are
-necessary:
-1) email notification
-2) text notification
-
-# Road Map
-
-~1) Make basic PyPi installation to get the idea of how it will work~
- ~- follow these instructions: https://packaging.python.org/tutorials/packaging-projects/~
-2 or 3) Setup text and email notification
- - follow these instructions: https://packaging.python.org/tutorials/packaging-projects/
-4) add advanced functionality.
- - create a list of phone provider gateways to output to user.
- - add other more advanced script features.
- - compile to binary(?) and set a CLI command to the PATH variable so a 
-   user can use a CMDLine shortcut.
-
