+++ python-port/Package/deleteme.md
@@ -1,5 +1,18 @@
 # Python Port of Teal Dulcet's sendmsg.sh Script
 
+# Road Map
+
+~1) Make basic PyPi installation to get the idea of how it will work~
+ ~- follow these instructions: https://packaging.python.org/tutorials/packaging-projects/~
+2 or 3) Setup text and email notification
+ - follow these instructions: https://packaging.python.org/tutorials/packaging-projects/
+4) add advanced functionality.
+ - create a list of phone provider gateways to output to user.
+ - add other more advanced script features.
+ - compile to binary(?) and set a CLI command to the PATH variable so a 
+   user can use a CMDLine shortcut.
+
+
 ## Libraries to use (Suggestion credit: Teal Dulcet)
 [email](https://docs.python.org/3/library/email.html)
 [smptlib](https://docs.python.org/3/library/smtplib.html)
