--- python-port/Package/realpython/.config2.ini
@@ -1,8 +1,0 @@
-[email]
-smtp = 
-username = 
-password = 
-
-[text]
-number = 
-carrier =
